+++ app/Services/Auth/UserRegistrar.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Services;
+namespace App\Services\Auth;
 
 use App\Exceptions\Auth\UserNotCreatedException;
 use App\Models\User;
